Add UpdatedAt and CreatedAt to models.Hobbit and models.NumericRecord
This commit adds UpdatedAt and CreatedAt to hobbit and record.
As those fields are sadly null after migration, we need to populate them
by ourselves. We either have the choice to create a empty / 0 date or
the current date. This pull request sets them to the current date.
However, this is subject to change.

Anyway, as we see manual migrations taking place in our codebase, we
should start to think about how we tackle migrations in general. Maybe
we should start storing more data, like the db schema version in the db
and execute migrations manually on startup? Maybe it even makes sense to
create a migrations directory containing .sql files for that. Not sure
:)