dependencies+store: Transition from vuex to pinia
This commit contains the first steps in order to convert our codebase
from vuex 4 to pinia, which is the recommended state solution for new
vue projects, especially in combination with the composition api. for
now, we still use it with the options api, where it also provides some
benefits such as better typescript support. the new state files are sooo
much shorter now. 🎉